feat: CQDG-794 rename utils index to utils doc
