+++ src/components/reports/DownloadFileManifestModal/index.tsx
@@ -10,7 +10,7 @@ import EnvVariables from 'helpers/EnvVariables';
 import TooMuchFilesAlert from 'components/reports/TooMuchFilesAlert';
 import { ReportType } from 'services/api/reports/models';
 import { fetchReport } from 'store/report/thunks';
-import { getDocLang } from 'utils';
+import { getDocLang } from 'utils/doc';
 
 import FilesTable from './FilesTable';
